refactored server actions
