+++ src/server/actions/achievement/index.ts
@@ -1,0 +1,6 @@
+// server actions don't like to be imported this way
+export { default as like } from "./like";
+export { default as unlike } from "./unlike";
+export { default as create } from "./create";
+
+export * from "./";
